Allocators: Add GP Allocator
The General Purpose Allocator is a DualRBTree allocator with two balanced binary search trees. These do use std::map and std::multimap for now. The General Purpose Allocator is meant to be used as a swift generic allocator. It has O(log(N)) time complexity and minimal overhead, which should help it perform better than or equal to the C Allocators.